+++ src/Reflection/Entity/ParameterEntity.php
@@ -18,8 +18,8 @@ use PhpDocMaker\Reflection\AbstractEntity;
 use PhpDocMaker\Reflection\AbstractMethodEntity;
 use PhpDocMaker\Reflection\Entity\FunctionEntity;
 use PhpDocMaker\Reflection\Entity\MethodEntity;
+use PhpDocMaker\Reflection\Entity\TagEntity;
 use PhpDocMaker\Reflection\Entity\Traits\GetTypeAsStringTrait;
-use phpDocumentor\Reflection\DocBlock\Tags\Param;
 use Roave\BetterReflection\Reflection\ReflectionMethod;
 use Roave\BetterReflection\Reflection\ReflectionParameter;
 
@@ -104,14 +104,14 @@ class ParameterEntity extends AbstractEntity
     {
         $params = $this->getDeclaringFunction()->getTagsByName('param');
 
-        //Takes the right parameter
-        $param = call_user_func(function () use ($params): ?Param {
-            return array_value_first(array_filter($params, function (Param $param) {
-                return $param->getVariableName() === $this->getName();
+        //Takes the right parameter tag
+        $param = call_user_func(function () use ($params): ?TagEntity {
+            return array_value_first(array_filter($params, function (TagEntity $tag) {
+                return $tag->getVariableName() === $this->getName();
             }));
         });
 
-        return $param ? $param->getDescription()->render() : '';
+        return $param ? $param->getDescription() : '';
     }
 
     /**