+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.x branch
 ## 1.0 branch
 ### 1.0.2-beta3
+* added `TagEntity`, which now represents the tags;
 * added `ClassEntity::getType()` method;
 * added `ClassEntity::createFromName()` static method;
 * added `TestCase::assertStringEqualsTemplate()` assertion method;